requirements.txt
decorator
docutils
psutil
robotframework
robotframework-selenium2library
selenium
Pillow
